# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Selection Task of Quick Deal",
        "version": "0.1"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Retrieve a list of tasks",
                "description": "Get a paginated list of tasks along with pagination information.",
                "operationId": "feed89770241fcb6af2325dafc9adfde",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Optional Query Param for Filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DRAFT",
                                "WAITING",
                                "PROCESSING",
                                "COMPLETED"
                            ],
                            "example": ""
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Optional Query Param for Filter",
                        "required": false,
                        "schema": {
                            "type": "date",
                            "example": "2024-02-29"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "body": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "DRAFT",
                                                            "WAITING",
                                                            "PROCESSING",
                                                            "COMPLETED"
                                                        ]
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Creating a Task",
                "operationId": "44140ecf3dfc93dfa6ae24da60ceb875",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "body": {
                                                "type": "text",
                                                "example": "Some long text for body because minimum characters length should be 32."
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "DRAFT",
                                                    "WAITING",
                                                    "PROCESSING",
                                                    "COMPLETED"
                                                ]
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-02-29"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some Title"
                                                },
                                                "body": {
                                                    "type": "text",
                                                    "example": "Some Body Description Here"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "DRAFT"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2020-02-03"
                                                },
                                                "updated_at": {
                                                    "type": "timestamp",
                                                    "example": "2024-02-28T23:34:07.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "timestamp",
                                                    "example": "2024-02-28T23:34:07.000000Z"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "data": {
                                            "properties": {
                                                "body": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The title must be at least 32 characters long."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}